+++ firmware/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Show random greetings to user
+
 ## 0.1.0 - 2024-10-30
 
 - Showcased on general meeting